refactor(PromissoryNotes): remove commented out code for entriesFinanacialReporting
